Split updater.rs into module
